api structure refactor
