+++ src/stores/conversationStore.ts
@@ -1,0 +1,587 @@
+import { ref, onUnmounted } from 'vue'
+import { defineStore, storeToRefs } from 'pinia'
+import type OpenAI from 'openai'
+import * as api from '../services/apiService'
+import { useGeneralStore } from './generalStore'
+import { useSettingsStore } from './settingsStore'
+import { executeFunction } from '../utils/functionCaller'
+import eventBus from '../utils/eventBus'
+
+export interface AppChatMessageContentPart {
+  type: 'app_text' | 'app_image_uri' | 'app_generated_image_path' | 'app_file'
+  text?: string
+  uri?: string
+  path?: string
+  absolutePathForOpening?: string
+  imageGenerationId?: string
+  isPartial?: boolean
+  partialIndex?: number
+  fileId?: string
+  fileName?: string
+}
+
+export interface ChatMessage {
+  local_id_temp?: string
+  api_message_id?: string
+  api_response_id?: string
+  role: 'user' | 'assistant' | 'system' | 'developer' | 'tool'
+  content: string | AppChatMessageContentPart[]
+  tool_call_id?: string
+  name?: string
+  tool_calls?: OpenAI.Responses.FunctionCall[]
+}
+
+interface RawMessageForSummarization {
+  role: string
+  text_content: string
+  created_at: string
+}
+
+export const useConversationStore = defineStore('conversation', () => {
+  const generalStore = useGeneralStore()
+  const settingsStore = useSettingsStore()
+  const {
+    setAudioState,
+    queueAudioForPlayback,
+    updateImageContentPartByGenerationId,
+  } = generalStore
+  const { isRecordingRequested, audioState, chatHistory } =
+    storeToRefs(generalStore)
+
+  const currentResponseId = ref<string | null>(null)
+  const currentConversationTurnId = ref<string | null>(null)
+  const isInitialized = ref<boolean>(false)
+  const availableModels = ref<OpenAI.Models.Model[]>([])
+  const isSummarizing = ref<boolean>(false)
+  const ttsAbortController = ref<AbortController | null>(null)
+  const llmAbortController = ref<AbortController | null>(null)
+  const ephemeralEmotionalContext = ref<string | null>(null)
+
+  const handleCancelTTS = () => {
+    if (ttsAbortController.value) {
+      console.log('[TTS Abort] Cancelling in-flight TTS request.')
+      ttsAbortController.value.abort()
+      ttsAbortController.value = null
+    }
+  }
+
+  const handleCancelLLMStream = () => {
+    if (llmAbortController.value) {
+      console.log('[LLM Abort] Cancelling in-flight LLM stream request.')
+      llmAbortController.value.abort()
+      llmAbortController.value = null
+    }
+  }
+
+  eventBus.on('cancel-tts', handleCancelTTS)
+  eventBus.on('cancel-llm-stream', handleCancelLLMStream)
+
+  onUnmounted(() => {
+    eventBus.off('cancel-tts', handleCancelTTS)
+    eventBus.off('cancel-llm-stream', handleCancelLLMStream)
+  })
+
+  const initialize = async (): Promise<boolean> => {
+    if (isInitialized.value) {
+      return true
+    }
+    if (!settingsStore.initialLoadAttempted) {
+      await settingsStore.loadSettings()
+    }
+
+    const essentialCheckPassed = settingsStore.areEssentialSettingsProvided
+    if (!essentialCheckPassed) {
+      generalStore.statusMessage =
+        'Error: Core settings (API Keys/Models/STT) not configured.'
+      isInitialized.value = false
+      return false
+    }
+
+    if (
+      availableModels.value.length === 0 &&
+      settingsStore.config.VITE_OPENAI_API_KEY
+    ) {
+      await fetchModels()
+    }
+    isInitialized.value = true
+    generalStore.statusMessage = 'Stand by'
+    generalStore.chatHistory = []
+    currentResponseId.value = null
+    currentConversationTurnId.value = `turn-${Date.now()}`
+    return true
+  }
+
+  const triggerConversationSummarization = async () => {
+    if (isSummarizing.value) return
+    isSummarizing.value = true
+
+    try {
+      const messagesResult = await window.ipcRenderer.invoke(
+        'summaries:get-recent-messages',
+        { limit: settingsStore.config.SUMMARIZATION_MESSAGE_COUNT }
+      )
+
+      if (messagesResult.success && messagesResult.data?.length > 0) {
+        const rawMessages = messagesResult.data as RawMessageForSummarization[]
+        const formattedMessages = rawMessages.map(m => ({
+          role: m.role,
+          content: m.text_content || '[content missing]',
+        }))
+
+        const [emotionalContext, factualSummary] = await Promise.all([
+          api.createContextAnalysisResponse(
+            formattedMessages,
+            settingsStore.config.SUMMARIZATION_MODEL
+          ),
+          api.createSummarizationResponse(
+            formattedMessages,
+            settingsStore.config.SUMMARIZATION_MODEL,
+            settingsStore.config.SUMMARIZATION_SYSTEM_PROMPT
+          ),
+        ])
+
+        if (emotionalContext) {
+          ephemeralEmotionalContext.value = emotionalContext
+        }
+
+        if (factualSummary) {
+          await window.ipcRenderer.invoke('summaries:save-summary', {
+            summaryText: factualSummary,
+            summarizedMessagesCount: rawMessages.length,
+          })
+        }
+      }
+    } catch (error) {
+      console.error('[Summarizer] Error during summarization:', error)
+    } finally {
+      isSummarizing.value = false
+    }
+  }
+
+  const buildApiInput = async (
+    isNewChain: boolean
+  ): Promise<OpenAI.Responses.Request.InputItemLike[]> => {
+    const historyToBuildFrom = [...chatHistory.value]
+    const apiInput: OpenAI.Responses.Request.InputItemLike[] = []
+    const recentHistory = historyToBuildFrom
+      .slice(0, settingsStore.config.MAX_HISTORY_MESSAGES_FOR_API)
+      .reverse()
+
+    const summaryResult = await window.ipcRenderer.invoke(
+      'summaries:get-latest-summary',
+      {}
+    )
+    if (summaryResult.success && summaryResult.data?.summary_text) {
+      const summaryContent = `[CONVERSATION_SUMMARY_START]\nContext from previous conversation segment:\n${summaryResult.data.summary_text}\n[CONVERSATION_SUMMARY_END]`
+      apiInput.push({
+        role: 'system',
+        content: [{ type: 'input_text', text: summaryContent }],
+      })
+    }
+
+    let lastUserMessageInFullHistoryId: string | null = null
+    for (let i = 0; i < historyToBuildFrom.length; i++) {
+      if (historyToBuildFrom[i].role === 'user') {
+        lastUserMessageInFullHistoryId =
+          historyToBuildFrom[i].local_id_temp || null
+        break
+      }
+    }
+
+    for (const msg of recentHistory) {
+      let apiItemPartial: any
+
+      if (msg.role === 'tool') {
+        apiItemPartial = {
+          type: 'function_call_output',
+          call_id: msg.tool_call_id || `unknown_call_id_${Date.now()}`,
+          output:
+            typeof msg.content === 'string'
+              ? msg.content
+              : JSON.stringify(msg.content),
+        }
+      } else if (msg.role === 'system') {
+        apiItemPartial = {
+          role: 'system',
+          content: [
+            {
+              type: 'input_text',
+              text:
+                typeof msg.content === 'string'
+                  ? msg.content
+                  : JSON.stringify(msg.content),
+            },
+          ],
+        }
+      } else {
+        const currentApiRole = msg.role as 'user' | 'assistant' | 'developer'
+        apiItemPartial = { role: currentApiRole, content: [] }
+        if (msg.name) apiItemPartial.name = msg.name
+
+        let messageContentParts: OpenAI.Responses.Request.ContentPartLike[] = []
+        if (typeof msg.content === 'string') {
+          const typeForPart =
+            currentApiRole === 'user' || currentApiRole === 'developer'
+              ? 'input_text'
+              : 'output_text'
+          messageContentParts = [{ type: typeForPart, text: msg.content }]
+        } else if (Array.isArray(msg.content)) {
+          messageContentParts = msg.content
+            .map((appPart: AppChatMessageContentPart) => {
+              if (appPart.type === 'app_text') {
+                const typeForPart =
+                  currentApiRole === 'user' || currentApiRole === 'developer'
+                    ? 'input_text'
+                    : 'output_text'
+                return { type: typeForPart, text: appPart.text || '' }
+              }
+              if (appPart.type === 'app_image_uri' && appPart.uri) {
+                if (
+                  currentApiRole === 'user' ||
+                  currentApiRole === 'developer'
+                ) {
+                  return {
+                    type: 'input_image',
+                    image_url: appPart.uri,
+                  } as OpenAI.Responses.Request.InputImagePart
+                }
+              }
+              return null
+            })
+            .filter(
+              p => p !== null
+            ) as OpenAI.Responses.Request.ContentPartLike[]
+        }
+        apiItemPartial.content =
+          messageContentParts.length > 0
+            ? messageContentParts
+            : [{ type: 'input_text', text: '' }]
+      }
+      apiInput.push(apiItemPartial)
+    }
+    return apiInput
+  }
+
+  async function _processStreamLogic(
+    stream: AsyncIterable<OpenAI.Responses.StreamEvent>,
+    placeholderTempId: string,
+    isContinuationAfterTool: boolean
+  ): Promise<{ streamEndedNormally: boolean }> {
+    let currentSentence = ''
+    let currentAssistantApiMessageId: string | null = null
+    const functionCallArgsBuffer: Record<string, string> = {}
+    let streamEndedNormally = true
+
+    const sendToTTS = async (text: string) => {
+      if (text.trim()) {
+        ttsAbortController.value = new AbortController()
+        try {
+          const ttsResponse = await api.ttsStream(
+            text,
+            ttsAbortController.value.signal
+          )
+          if (
+            queueAudioForPlayback(ttsResponse) &&
+            audioState.value !== 'SPEAKING'
+          ) {
+            setAudioState('SPEAKING')
+          }
+        } catch (error: any) {
+          if (error.name !== 'AbortError') {
+            console.error('TTS stream creation failed:', error)
+          }
+        }
+      }
+    }
+
+    try {
+      for await (const event of stream) {
+        if (event.type === 'response.created') {
+          currentResponseId.value = event.response.id
+          generalStore.updateMessageApiResponseIdByTempId(
+            placeholderTempId,
+            event.response.id
+          )
+        }
+
+        if (
+          event.type === 'response.output_item.added' ||
+          event.type === 'response.output_item.updated'
+        ) {
+          if (
+            event.item.type === 'message' &&
+            event.item.role === 'assistant'
+          ) {
+            currentAssistantApiMessageId = event.item.id
+            if (currentAssistantApiMessageId) {
+              generalStore.updateMessageApiIdByTempId(
+                placeholderTempId,
+                currentAssistantApiMessageId
+              )
+            }
+          }
+        }
+
+        if (
+          event.type === 'response.output_text.delta' &&
+          event.item_id === currentAssistantApiMessageId
+        ) {
+          const textChunk = event.delta || ''
+          currentSentence += textChunk
+          generalStore.appendMessageDeltaByTempId(placeholderTempId, textChunk)
+          if (textChunk.match(/[.!?]\s*$/) || textChunk.includes('\n')) {
+            await sendToTTS(currentSentence)
+            currentSentence = ''
+          }
+        }
+
+        if (event.type === 'response.function_call_arguments.delta') {
+          const itemId = event.item_id
+          functionCallArgsBuffer[itemId] =
+            (functionCallArgsBuffer[itemId] || '') + (event.delta || '')
+        }
+
+        if (
+          event.type === 'response.output_item.done' &&
+          event.item.type === 'function_call'
+        ) {
+          const functionCallPayload =
+            event.item as OpenAI.Responses.FunctionCall
+          let args = {}
+          try {
+            args = JSON.parse(
+              functionCallArgsBuffer[functionCallPayload.id] || '{}'
+            )
+          } catch (e) {
+            console.error('Error parsing function call arguments:', e)
+          }
+          functionCallPayload.arguments = args
+          generalStore.addToolCallToMessageByTempId(
+            placeholderTempId,
+            functionCallPayload
+          )
+          await handleToolCall(functionCallPayload, currentResponseId.value)
+          return { streamEndedNormally: true }
+        }
+
+        if (event.type === 'error') {
+          streamEndedNormally = false
+          console.error('Streaming error event:', event.error)
+          break
+        }
+      }
+      await sendToTTS(currentSentence)
+    } catch (error: any) {
+      if (error.name !== 'AbortError') {
+        streamEndedNormally = false
+        console.error('Error processing stream:', error)
+      }
+    } finally {
+      if (streamEndedNormally) {
+        const finalizationInterval = setInterval(() => {
+          if (generalStore.audioPlayer && !generalStore.audioPlayer.paused)
+            return
+          if (generalStore.audioQueue.length > 0) return
+
+          if (
+            audioState.value === 'SPEAKING' ||
+            audioState.value === 'WAITING_FOR_RESPONSE'
+          ) {
+            setAudioState(isRecordingRequested.value ? 'LISTENING' : 'IDLE')
+          }
+          clearInterval(finalizationInterval)
+          if (!isContinuationAfterTool) {
+            triggerConversationSummarization()
+          }
+        }, 250)
+      }
+    }
+    return { streamEndedNormally }
+  }
+
+  async function handleToolCall(
+    toolCall: OpenAI.Responses.FunctionCall,
+    originalResponseIdForTool: string | null
+  ) {
+    const functionName = toolCall.name
+    const functionArgs = toolCall.arguments as object
+
+    const toolStatusMessage = getToolStatusMessage(functionName)
+    generalStore.addMessageToHistory({
+      role: 'system',
+      content: [{ type: 'app_text', text: toolStatusMessage }],
+    })
+
+    const resultString = await executeFunction(functionName, functionArgs)
+
+    generalStore.addMessageToHistory({
+      role: 'tool',
+      tool_call_id: toolCall.call_id,
+      name: functionName,
+      content: resultString,
+    })
+
+    const isNewChainAfterTool = currentResponseId.value === null
+    const nextApiInput = await buildApiInput(isNewChainAfterTool)
+
+    const afterToolPlaceholder: ChatMessage = {
+      role: 'assistant',
+      content: [{ type: 'app_text', text: '' }],
+    }
+    const afterToolPlaceholderTempId =
+      generalStore.addMessageToHistory(afterToolPlaceholder)
+
+    llmAbortController.value = new AbortController()
+    try {
+      const continuedStream = await api.createOpenAIResponse(
+        nextApiInput,
+        originalResponseIdForTool,
+        true,
+        settingsStore.config.assistantSystemPrompt,
+        llmAbortController.value.signal
+      )
+      await _processStreamLogic(
+        continuedStream,
+        afterToolPlaceholderTempId,
+        true
+      )
+    } catch (error: any) {
+      if (error.name !== 'AbortError') {
+        console.error('Error in continued stream after tool call:', error)
+        generalStore.updateMessageContentByTempId(
+          afterToolPlaceholderTempId,
+          `Error: ${error.message}`
+        )
+        setAudioState(isRecordingRequested.value ? 'LISTENING' : 'IDLE')
+      }
+    }
+  }
+
+  const chat = async () => {
+    if (!isInitialized.value) {
+      console.warn('Conversation store not initialized.')
+      return
+    }
+    currentConversationTurnId.value = `turn-${Date.now()}`
+    setAudioState('WAITING_FOR_RESPONSE')
+
+    const isNewChain = currentResponseId.value === null
+    const constructedApiInput = await buildApiInput(isNewChain)
+
+    let finalInstructions = settingsStore.config.assistantSystemPrompt
+    const latestUserMessageContent = chatHistory.value.find(
+      m => m.role === 'user'
+    )?.content
+    if (
+      typeof latestUserMessageContent === 'string' &&
+      latestUserMessageContent
+    ) {
+      const relevantThoughts = await api.retrieveRelevantThoughtsForPrompt(
+        latestUserMessageContent
+      )
+      if (relevantThoughts.length > 0) {
+        const thoughtsBlock =
+          'Relevant thoughts from past conversation:\n' +
+          relevantThoughts.map(t => `- ${t}`).join('\n')
+        finalInstructions = `${thoughtsBlock}\n\n---\n\n${finalInstructions}`
+      }
+    }
+
+    if (ephemeralEmotionalContext.value) {
+      finalInstructions = `[System Note: User's tone was: "${ephemeralEmotionalContext.value}"]\n${finalInstructions}`
+      ephemeralEmotionalContext.value = null
+    }
+
+    const assistantMessagePlaceholder: ChatMessage = {
+      role: 'assistant',
+      content: [{ type: 'app_text', text: '' }],
+    }
+    const placeholderTempId = generalStore.addMessageToHistory(
+      assistantMessagePlaceholder
+    )
+
+    try {
+      llmAbortController.value = new AbortController()
+      const streamResult = await api.createOpenAIResponse(
+        constructedApiInput,
+        currentResponseId.value,
+        true,
+        finalInstructions,
+        llmAbortController.value.signal
+      )
+      await _processStreamLogic(streamResult, placeholderTempId, false)
+    } catch (error: any) {
+      if (error.name !== 'AbortError') {
+        console.error('Error starting OpenAI response stream:', error)
+        generalStore.updateMessageContentByTempId(
+          placeholderTempId,
+          `Error: ${error.message}`
+        )
+        setAudioState(isRecordingRequested.value ? 'LISTENING' : 'IDLE')
+      }
+    }
+  }
+
+  const transcribeAudioMessage = async (
+    audioArrayBuffer: ArrayBuffer
+  ): Promise<string> => {
+    const { sttProvider } = settingsStore.config
+    try {
+      if (sttProvider === 'openai') {
+        return await api.transcribeWithOpenAI(audioArrayBuffer)
+      } else {
+        return await api.transcribeWithGroq(audioArrayBuffer)
+      }
+    } catch (error: any) {
+      generalStore.statusMessage = 'Error: Transcription failed'
+      console.error('Transcription service error:', error)
+      return ''
+    }
+  }
+
+  async function fetchModels() {
+    if (!settingsStore.config.VITE_OPENAI_API_KEY) {
+      console.warn('Cannot fetch models: OpenAI API Key is missing.')
+      return
+    }
+    try {
+      availableModels.value = await api.fetchOpenAIModels()
+    } catch (error: any) {
+      console.error('Failed to fetch models:', error.message)
+      generalStore.statusMessage = `Error: Could not fetch AI models.`
+      availableModels.value = []
+    }
+  }
+
+  function getToolStatusMessage(toolName: string): string {
+    const messages: Record<string, string> = {
+      get_current_datetime: '🕒 Looking at the clock...',
+      open_path: '📂 Opening that for you...',
+      manage_clipboard: '📋 Working with your clipboard...',
+      search_torrents: '🧲 Looking for torrents...',
+      add_torrent_to_qb: '🚀 Starting your download...',
+      save_memory: '🧠 Got it, remembering that...',
+      recall_memories: '🧠 Let me think back...',
+      delete_memory: '🗑️ Forgetting that now...',
+      get_calendar_events: '🗓️ Fetching your schedule...',
+      create_calendar_event: '🗓️ Adding to your calendar...',
+      update_calendar_event: '🗓️ Updating your calendar...',
+      delete_calendar_event: '🗑️ Removing from your calendar...',
+      get_unread_emails: '📧 Looking for unread emails...',
+      search_emails: '📧 Searching emails...',
+      get_email_content: '📧 Reading email content...',
+    }
+    return messages[toolName] || `⚙️ Using tool: ${toolName}...`
+  }
+
+  return {
+    isInitialized,
+    availableModels,
+    initialize,
+    chat,
+    transcribeAudioMessage,
+    fetchModels,
+    currentResponseId,
+    triggerConversationSummarization,
+  }
+})
